アイソメトリック座標変換をutilに移動
